experiment: see if it makes sense to have generic form html
this test proves that with WTF_Forms, python and Bootstrap, we can
get away wiht have a generic form html, and just pass the form
object created from the forms.py.  In this case, I added a title
attribute that can be used to make the 'title' that's passed to
the html.  Because title is used in the base html, it doesn't
allow me to no pass it.